Pull out response types into separate`types.ts` to avoid circular imports
